--- Team-B/src/test/java/org/springframework/samples/petclinic/vet/SpecialtyTest.java
@@ -1,7 +1,0 @@
-package org.springframework.samples.petclinic.vet;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class SpecialtyTest {
-
-}
